Disentangle circular dependency
