+++ server/models/services/AtVersionService.js
@@ -1,0 +1,52 @@
+const ModelService = require('./ModelService');
+const { AtVersion } = require('..');
+const { Op } = require('sequelize');
+const { AT_VERSION_ATTRIBUTES, AT_ATTRIBUTES } = require('./helpers');
+
+/**
+ * @param {object} options
+ * @param {string|any} options.search - use this to combine with {@param filter} to be passed to Sequelize's where clause
+ * @param {object} options.where - use this define conditions to be passed to Sequelize's where clause
+ * @param {string[]} options.atVersionAttributes  - AtVersion attributes to be returned in the result
+ * @param {string[]} options.atAttributes  - At attributes to be returned in the result
+ * @param {object} options.pagination - pagination options for query
+ * @param {number} options.pagination.page - page to be queried in the pagination result (affected by {@param pagination.enablePagination})
+ * @param {number} options.pagination.limit - amount of results to be returned per page (affected by {@param pagination.enablePagination})
+ * @param {string[][]} options.pagination.order- expects a Sequelize structured input dataset for sorting the Sequelize Model results (NOT affected by {@param pagination.enablePagination}). See {@link https://sequelize.org/v5/manual/querying.html#ordering} and {@example [ [ 'username', 'DESC' ], [..., ...], ... ]}
+ * @param {boolean} options.pagination.enablePagination - use to enable pagination for a query result as well useful values. Data for all items matching query if not enabled
+ * @param {*} options.transaction - Sequelize transaction
+ * @returns {Promise<*>}
+ */
+const getAtVersions = async ({
+  search,
+  where = {},
+  atVersionAttributes = AT_VERSION_ATTRIBUTES,
+  atAttributes = AT_ATTRIBUTES,
+  pagination = {},
+  transaction
+}) => {
+  // search and filtering options
+  const searchQuery = search ? `%${search}%` : '';
+  if (searchQuery) where = { ...where, name: { [Op.iLike]: searchQuery } };
+
+  return ModelService.get(AtVersion, {
+    where,
+    attributes: atVersionAttributes,
+    include: [atAssociation(atAttributes)],
+    pagination,
+    transaction
+  });
+};
+
+/**
+ * @param atAttributes - At attributes to be returned in the result
+ * @returns {{association: string, attributes: string[]}}
+ */
+const atAssociation = atAttributes => ({
+  association: 'at',
+  attributes: atAttributes
+});
+
+module.exports = {
+  getAtVersions
+};
