+++ server/tests/models/services/ModelService.test.js
@@ -5,6 +5,7 @@ const {
   USER_ROLES_ATTRIBUTES
 } = require('../../../models/services/helpers');
 const AtService = require('../../../models/services/AtService');
+const AtVersionService = require('../../../models/services/AtVersionService');
 const BrowserService = require('../../../models/services/BrowserService');
 const dbCleaner = require('../../util/db-cleaner');
 
@@ -157,7 +158,7 @@ describe('ModelService', () => {
       const results = await ModelService.nestedGetOrCreate({
         operations: [
           {
-            get: AtService.getAtVersions,
+            get: AtVersionService.getAtVersions,
             create: AtService.createAtVersion,
             values: { atId: _atId, name: _atVersion },
             returnAttributes: {}
